+++ tests/test_trial.py
@@ -112,7 +112,7 @@ class TestTrial:
             [m for m in trial._gather_all_messages("test_source")]
 
     def test_add_feedback(self, trial, unittest_case, data_pb2):
-        trial.add_feedback(
+        trial.add_reward(
             value=12,
             confidence=1.0,
             to=["agent_1", "agent_2"],
@@ -120,7 +120,7 @@ class TestTrial:
             user_data=data_pb2.MyFeedbackUserData(a_bool=False, a_float=3.0)
         )
 
-        trial.add_feedback(
+        trial.add_reward(
             value=-2,
             confidence=0.,
             to=["agent_3"],
@@ -128,23 +128,26 @@ class TestTrial:
             user_data=data_pb2.MyFeedbackUserData(a_bool=True, a_float=3.0)
         )
 
-        feedbacks_as_tuples = []
-        for f in trial._gather_all_feedback():
+        rewards_as_tuples = []
+        for rew in trial._gather_all_rewards():
             user_data = data_pb2.MyFeedbackUserData()
-            assert f.user_data.Unpack(user_data)
-            feedbacks_as_tuples.append((f.actor_name, f.value, f.confidence, f.tick_id, user_data))
+            assert len(rew.sources) == 1
+            src = rew.sources[0]
+            assert src.user_data.Unpack(user_data)
+            rewards_as_tuples.append(
+                (rew.receiver_name, src.value, src.confidence, rew.tick_id, user_data))
 
-        assert len(feedbacks_as_tuples) == 3
-        unittest_case.assertCountEqual(feedbacks_as_tuples, [
+        assert len(rewards_as_tuples) == 3
+        unittest_case.assertCountEqual(rewards_as_tuples, [
             ("agent_1", 12, 1., 666, data_pb2.MyFeedbackUserData(a_bool=False, a_float=3.0)),
             ("agent_2", 12, 1., 666, data_pb2.MyFeedbackUserData(a_bool=False, a_float=3.0)),
             ("agent_3", -2, 0., 667, data_pb2.MyFeedbackUserData(a_bool=True, a_float=3.0))
         ])
 
-        assert len([m for m in trial._gather_all_feedback()]) == 0
+        assert len([m for m in trial._gather_all_rewards()]) == 0
 
     def test_add_feedback_bad_user_data(self, trial):
-        trial.add_feedback(
+        trial.add_reward(
             value=12,
             confidence=1.0,
             to=["agent_1", "agent_2"],
@@ -154,4 +157,4 @@ class TestTrial:
 
         # Maybe this should be caught earlier
         with pytest.raises(AttributeError):
-            [f for f in trial._gather_all_feedback()]
+            [f for f in trial._gather_all_rewards()]