+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Changed
+
+- Removal of "feedback" in rewards.
+
 ## v1.0.0-alpha5 - 2021-01-28
 
 ### Added